Move `data/provider.ts` to `data/dataProvider.ts`
